refactor(indexer): get rid of tx_changes_processor
