+++ crates/iota-indexer/src/handlers/mod.rs
@@ -4,4 +4,3 @@
 
 pub mod optimistic_pruner;
 pub mod pruner;
-pub mod tx_processor;
